requirements.txt
aiogram==3.19.0
aiohttp
python-dotenv

